Lint

+++ src/app/api/users/profile/delete/route.ts
@@ -4,16 +4,15 @@ import { db } from "../../../../../db";
 import { getServerSession } from "next-auth/next";
 import { authOptions } from "../../../../../utils/authOptions";
 import { eq } from "drizzle-orm";
+import { Session } from "next-auth";
 
 export async function DELETE() {
+  const session = (await getServerSession(authOptions)) as Session;
 
-    const session = await getServerSession(authOptions);
-
-    try {
-        await db.delete(profiles)
-                .where(eq(profiles.user_id, session.user.id));
-        return NextResponse.json("Profile sucsessfully deleted");
-    } catch (error) {
-        return NextResponse.json(error);
-    }
+  try {
+    await db.delete(profiles).where(eq(profiles.user_id, session.user.id));
+    return NextResponse.json("Profile sucsessfully deleted");
+  } catch (error) {
+    return NextResponse.json(error);
+  }
 }
